Refactor: Separated project into multiple files
refactor
