+++ src/leaderboard/Models/Contestant.tsx
@@ -1,7 +1,6 @@
 import { RoundData } from "./RoundData"
 
 export interface Contestant {
-
    id: string
    name: string
    points: number
